requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
requests
google-auth
python-dotenv
